// apps/api/src/modules/plugins/impl/rust-plugin/utils/types.ts
/*
 * Copyright (C) 2026 Rubén Santibáñez Acosta
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published
 * by the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 * GNU Affero General Public License for more details.
 */

export type { PluginContext } from '../../../../../plugins-core/plugin.interface';
export interface Repository {
  id: string;
  name: string;
  type: string;
  manager: string;
  config?: any;
}